Add Calico BGP options with allowed_address_pairs
Calico in the Helm chart defaults to VXLAN encapsulation which
may have performance problems with vxlan offloading disabled[1].

Using BGP and no encapsulation for the CNI has much better
performance when vxlan offloading is disabled, but requires
allowed_address_pairs to be set by CAPO on Neutron ports, and
the full mesh overhead becomes a burden with clusters >50 nodes.

So, this option will suit some users, but not all.

Depends-on: https://github.com/azimuth-cloud/capi-helm-charts/pull/388

[1] VXLAN offloading may be disabled on the hardware, or by Calico
    for kernels > 5.7.
    See "This release re-enables VXLAN checksum offload for kernels > 5.7"
    in https://github.com/projectcalico/calico/blob/v3.28.0/release-notes/v3.28.0-release-notes.md

Change-Id: Ib01f257ff6e600863f7690a22a7384f5c58b60bd