Improve default min/max node count behaviour when autoscaling is enabled
This change fixes an issue where creating a cluster with the 2 or more nodes in the default worker group and with `auto_scaling_enabled=true` but with neither `{min,max}_node_count` labels set would lead to a situation where the default worker group would only have 1 node instead of 2. The reason for this seems to be that the `min_node_count` and `max_node_count` values which are passed through to CAPI take priority over the initial `node_count` field and the min count currently defaults to 1. In these cases, I think the expected user behaviour is to have `node_count` == `min_node_count` == `max_node_count` and for autoscaling to be disabled on the default node group.

Change-Id: I65e65bf64c97cf5ab659d6d7d158c32692a63348