Autoscale and max/min nodes now calculated from nodegroup fields
When creating a Cluster, labels are used to provide min_node_count
and max_node_count as the API provides no way to set these fields
directly on the nodegroup.

However, as labels cannot be changed and certain fields can be, this
change allows fields to be used over labels when they are set in
future.

Closes-bug: https://bugs.launchpad.net/magnum/+bug/2111863

Change-Id: I5dc69d6916c76efcf54c371088de59aa40fc4fed